+++ examples/guide/01_integrator.py
@@ -5,7 +5,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from modypy.model import System, State, signal_function
-from modypy.simulation import Simulator
+from modypy.simulation import Simulator, SimulationResult
 
 # Create a new system
 system = System()
@@ -23,20 +23,15 @@ integrator_state = State(system,
                          derivative_function=cosine_input)
 
 # Set up a simulation
-simulator = Simulator(system,
-                      start_time=0.0)
+simulator = Simulator(system, start_time=0.0)
 
-# Run the simulation for 10s
-simulator.run_until(time_boundary=10.0)
+# Run the simulation for 10s and capture the result
+result = SimulationResult(system, simulator.run_until(time_boundary=10.0))
 
 # Plot the result
 input_line, integrator_line = \
-    plt.plot(simulator.result.time,
-             cosine_input(simulator.result),
-             "r",
-             simulator.result.time,
-             integrator_state(simulator.result)[0],
-             "g")
+    plt.plot(result.time, cosine_input(result), "r",
+             result.time, integrator_state(result)[0], "g")
 plt.legend((input_line, integrator_line), ("Input", "Integrator State"))
 plt.title("Integrator")
 plt.xlabel("Time")