+++ examples/guide/02_pendulum.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 
 from modypy.blocks.linear import integrator
 from modypy.model import System, State
-from modypy.simulation import Simulator
+from modypy.simulation import Simulator, SimulationResult
 
 # Define the system parameters
 LENGTH = 1.0
@@ -34,18 +34,14 @@ omega = State(system,
 # Create the alpha state
 alpha = integrator(system, input_signal=omega, initial_condition=ALPHA_0)
 
-# Run a simulation
+# Run a simulation and capture the result
 simulator = Simulator(system, start_time=0.0)
-simulator.run_until(time_boundary=10.0)
+result = SimulationResult(system, simulator.run_until(time_boundary=10.0))
 
 # Plot the result
 alpha_line, omega_line = \
-    plt.plot(simulator.result.time,
-             alpha(simulator.result)[0],
-             "r",
-             simulator.result.time,
-             omega(simulator.result)[0],
-             "g")
+    plt.plot(result.time, alpha(result)[0], "r",
+             result.time, omega(result)[0], "g")
 plt.legend((alpha_line, omega_line), ("Alpha", "Omega"))
 plt.title("Pendulum")
 plt.xlabel("Time")