+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   leads to signals/ports not being explicitly registered in the system anymore.
 - The results in ``SimulatorResult`` are now shaped differently: The last index
   represents the sample number (#21)
+- The ``run_until`` method now returns a generator for the simulation sample
+  points. The values returned are system states. They can be cached using a
+  ``SimulatorResult``.
+  This allows for more flexibility on working with simulation data (e.g., 
+  when doing continuous simulation).
 
 ## [2.1.0] - 2021-03-07
 ### Added